globals.py
import os
from dotenv import load_dotenv
load_dotenv()

REDIS_HOSTNAME = os.getenv("REDIS_HOSTNAME")
REDIS_PORT = os.getenv("REDIS_PORT")
REDIS_PASSWORD = os.getenv("REDIS_PASSWORD")
MONGO_HOST = os.getenv("MONGO_HOST")
DB = os.getenv('MONGO_DB')
PORT = os.getenv('MONGO_PORT')
MONGO_USER = os.getenv('MONGO_USER')
MONGO_PASSWORD = os.getenv('MONGO_PASSWORD')

embeddings = []


def add_to_embeddings(username, encoding):
    interm = {
        "name": username,
        "encoding": encoding
    }
    embeddings.append(interm)

