requirements.txt
fastapi
python-multipart
mongoengine
python-dotenv
face_recognition
uvicorn
blinker